Use official tauri version
