+++ Cargo.toml
@@ -35,19 +35,22 @@ frida-gum = { version = "0.13.2", features = [
 ] }
 
 
+# tauri_fuzz_tools = { path = "./tauri_fuzz_tools" }
+tauri_fuzz_tools = { git = "ssh://git@github.com/crabnebula-dev/tauri-fuzzer.git", branch = "fined_grained_interception" }
 [workspace.dependencies.tauri]
 # path = "../tauri/core/tauri"
-git = "ssh://git@github.com/crabnebula-dev/tauri.git"
-branch = "feat/fuzz"
+# git = "ssh://git@github.com/crabnebula-dev/tauri.git"
+git = "ssh://git@github.com/tauri-apps/tauri.git"
+branch = "1.x"
 default-features = false
 features = ["test"]
 
 [workspace.dependencies.tauri-runtime]
 # path = "../tauri/core/tauri-runtime"
-git = "ssh://git@github.com/crabnebula-dev/tauri.git"
-branch = "feat/fuzz"
+git = "ssh://git@github.com/tauri-apps/tauri.git"
+branch = "1.x"
 
 [workspace.dependencies.tauri-build]
 # path = "../tauri/core/tauri-build"
-git = "ssh://git@github.com/crabnebula-dev/tauri.git"
-branch = "feat/fuzz"
+git = "ssh://git@github.com/tauri-apps/tauri.git"
+branch = "1.x"
