+++ tauri_fuzz_tools/src/fuzz_utils.rs
@@ -1,116 +1,25 @@
 use serde::de::DeserializeOwned;
-use serde_json::Value as JsonValue;
 use std::collections::HashMap;
 use std::fmt::Debug;
 use tauri::api::ipc::CallbackFn;
 use tauri::test::MockRuntime;
-use tauri::test::{Ipc, IpcKey};
 use tauri::App;
-use tauri::AppHandle;
 use tauri::Builder;
 use tauri::InvokePayload;
 use tauri::Manager;
-use tauri::RunEvent;
 
 /// Minimal builder for a Tauri application using the `MockRuntime`
 pub fn mock_builder_minimal() -> Builder<MockRuntime> {
     Builder::<MockRuntime>::new()
 }
 
-/// Invoke a Tauri command given an `InvokePayload`.
-///
-/// The application processes the command and stops.
-///
-/// # Examples
-///
-/// ```rust
-///
-/// #[tauri::command]
-/// fn ping() -> &'static str {
-///   "pong"
-/// }
-///
-/// use tauri_fuzz_tools::*;
-/// use tauri::test::*;
-/// fn create_app<R: tauri::Runtime>(mut builder: tauri::Builder<R>) -> tauri::App<R> {
-///   builder
-///     .invoke_handler(tauri::generate_handler![ping])
-///     // remove the string argument on your app
-///     .build(mock_context(noop_assets()))
-///     .expect("failed to build app")
-/// }
-///
-/// let app = create_app(mock_builder());
-/// let payload = create_invoke_payload("ping".into(), CommandArgs::new());
-/// let res: Result<String, String> = invoke_command_and_stop(app, payload);
-/// assert_eq!(res, Ok("pong".into()));
-/// ```
-pub fn invoke_command_and_stop<T: DeserializeOwned + Debug>(
-    mut app: App<MockRuntime>,
-    payload: InvokePayload,
-) -> Result<T, T> {
-    let (tx, rx) = std::sync::mpsc::channel();
-    let w = app.get_window("main").expect("Could not get main window");
-    let callback = payload.callback;
-    let error = payload.error;
-
-    let ipc = w.state::<Ipc>();
-    ipc.insert_ipc(IpcKey { callback, error }, tx);
-    w.on_message(payload).unwrap();
-
-    app.run_iteration();
-    let res: Result<JsonValue, JsonValue> =
-        rx.recv().expect("Failed to receive result from command");
-    res.map(|v| serde_json::from_value(v).unwrap())
-        .map_err(|e| serde_json::from_value(e).unwrap())
-}
-
-/// Invoke a Tauri command given an `InvokePayload`.
-///
-/// A callback can be provided to interact with future events.
-///
-/// # Examples
-///
-/// ```rust
-/// #[tauri::command]
-/// fn ping() -> &'static str {
-///   "pong"
-/// }
-///
-/// use tauri_fuzz_tools::*;
-/// use tauri::test::*;
-/// fn create_app<R: tauri::Runtime>(mut builder: tauri::Builder<R>) -> tauri::App<R> {
-///   builder
-///     .invoke_handler(tauri::generate_handler![ping])
-///     // remove the string argument on your app
-///     .build(mock_context(noop_assets()))
-///     .expect("failed to build app")
-/// }
-///
-/// let app = create_app(mock_builder());
-/// let payload = create_invoke_payload("ping".into(), CommandArgs::new());
-/// // Exit the application after processing the command
-/// invoke_command(app, payload, move |app_handle, event| {
-///   match event {
-///      tauri::RunEvent::Ready => app_handle.exit(0),
-///      _ => (),
-///   }
-/// });
-/// ```
-pub fn invoke_command<F: FnMut(&AppHandle<MockRuntime>, RunEvent) + 'static>(
+/// Invoke a command and get the Tauri command returned value
+pub fn invoke_command<T: DeserializeOwned + Debug>(
     app: App<MockRuntime>,
     payload: InvokePayload,
-    event_callback: F,
-) {
+) -> Result<T, T> {
     let w = app.get_window("main").expect("Could not get main window");
-
-    let (tx, _rx) = std::sync::mpsc::channel();
-    let callback = payload.callback;
-    let error = payload.error;
-    let ipc = w.state::<Ipc>();
-    ipc.insert_ipc(IpcKey { callback, error }, tx);
-    w.on_message(payload).unwrap();
-    app.run(event_callback);
+    tauri::test::get_ipc_response::<T>(&w, payload)
 }
 
 /// Invoke a command but does not try to get the command return value
@@ -167,17 +76,31 @@ mod test {
 
     #[allow(dead_code)]
     #[tauri::command]
-    fn test_command() {}
+    fn test_command() -> String {
+        String::from("foo")
+    }
 
-    // #[test]
-    #[allow(dead_code)]
-    fn invoke_1_command() {
+    #[test]
+    fn test_invoke_command() {
         let app = mock_builder()
             .invoke_handler(tauri::generate_handler![test_command])
             .build(mock_context(noop_assets()))
             .unwrap();
         let payload = create_invoke_payload("test_command", CommandArgs::new());
-        let res = invoke_command_and_stop::<String>(app, payload);
+        let res = invoke_command::<String>(app, payload);
         assert!(res.is_ok());
+        assert_eq!(&res.unwrap(), "foo");
+    }
+
+    #[test]
+    fn test_invoke_minimal() {
+        let app = mock_builder_minimal()
+            .invoke_handler(tauri::generate_handler![test_command])
+            .build(mock_context(noop_assets()))
+            .unwrap();
+        let payload = create_invoke_payload("test_command", CommandArgs::new());
+        invoke_command_minimal(app, payload);
+        // The goal is just to reach this point
+        assert!(true);
     }
 }
